+++ tests/Api/BookingTest.php
@@ -76,7 +76,7 @@ class BookingTest extends AbstractBaseApiTestCase
 
     public function testBookingWebform(): void
     {
-        $this->messenger('async')->queue()->assertEmpty();
+        $this->transport('async')->queue()->assertEmpty();
 
         $client = $this->getAuthenticatedClient();
 
@@ -102,15 +102,15 @@ class BookingTest extends AbstractBaseApiTestCase
         $this->assertResponseStatusCodeSame(201);
 
         /** @var WebformSubmitMessage $message */
-        $message = $this->messenger('async')->queue()->first(WebformSubmitMessage::class)->getMessage();
+        $message = $this->transport('async')->queue()->first(WebformSubmitMessage::class)->getMessage();
         $this->assertEquals('booking', $message->getWebformId());
         $this->assertEquals('795f5a1c-a0ac-4f8a-8834-bb71fca8585d', $message->getSubmissionUuid());
         $this->assertEquals('https://bookaarhus.local.itkdev.dk', $message->getSender());
         $this->assertEquals('https://bookaarhus.local.itkdev.dk/webform_rest/booking/submission/123123123', $message->getSubmissionUrl());
         $this->assertEquals(1, $message->getApiKeyUserId());
 
-        $this->messenger('async')->queue()->assertCount(1);
-        $this->messenger('async')->queue()->assertContains(WebformSubmitMessage::class);
+        $this->transport('async')->queue()->assertCount(1);
+        $this->transport('async')->queue()->assertContains(WebformSubmitMessage::class);
     }
 
     /**
@@ -118,7 +118,7 @@ class BookingTest extends AbstractBaseApiTestCase
      */
     public function testWebformSubmitMessageHandler(): void
     {
-        $this->messenger('async')->queue()->assertEmpty();
+        $this->transport('async')->queue()->assertEmpty();
 
         $webformServiceMock = $this->getMockBuilder(WebformService::class)
             ->onlyMethods(['getWebformSubmission', 'getData'])
@@ -208,8 +208,8 @@ class BookingTest extends AbstractBaseApiTestCase
             $testUser->getId()
         ));
 
-        $this->messenger('async')->queue()->assertContains(CreateBookingMessage::class);
-        $this->messenger('async')->queue()->assertCount(1);
+        $this->transport('async')->queue()->assertContains(CreateBookingMessage::class);
+        $this->transport('async')->queue()->assertCount(1);
     }
 
     public function testCreateBookingHandler(): void
@@ -288,7 +288,7 @@ class BookingTest extends AbstractBaseApiTestCase
 
     public function testInvalidBookingWebform(): void
     {
-        $this->messenger('async')->queue()->assertEmpty();
+        $this->transport('async')->queue()->assertEmpty();
 
         $client = $this->getAuthenticatedClient();
 
@@ -309,6 +309,6 @@ class BookingTest extends AbstractBaseApiTestCase
 
         $this->assertResponseStatusCodeSame(400);
 
-        $this->messenger('async')->queue()->assertCount(0);
+        $this->transport('async')->queue()->assertCount(0);
     }
 }